+++ __init__.py
@@ -4,7 +4,7 @@ bl_info = {
     "name": "Storytools - Storyboard Tools",
     "description": "Set of tools for Storyboarding",
     "author": "Samuel Bernou",
-    "version": (1, 12, 0),
+    "version": (1, 12, 1),
     "blender": (3, 3, 0),
     "location": "View3D",
     "warning": "",